requirements.txt
aioredis==1.3.1
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.1
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.2
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
django-redis-cache==2.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.5.0
incremental==17.5.0
isort==4.3.21
kombu==4.6.8
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
msgpack==0.6.2
mysqlclient==1.4.6
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
python-dotenv==0.13.0
pytz==2019.3
redis==3.4.1
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
typed-ast==1.4.1
urllib3==1.25.8
vine==1.3.0
virtualenv==20.0.4
whitenoise==5.0.1
wrapt==1.11.2
zipp==3.0.0
zope.interface==4.7.1
